Move the Preferences key to an enum folder
